backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
anyio==4.6.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.9.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
common==0.1.2
configobj==5.0.9
configparser==7.1.0
contourpy==1.2.0
credentials==1.1
cryptography==43.0.1
cycler==0.12.1
defusedxml==0.7.1
Django==5.0.7
django-browser-reload==1.13.0
django-cors-headers==4.4.0
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
et-xmlfile==1.1.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.50.0
frontend==0.0.3
frozenlist==1.4.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
ipwhois==1.3.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==3.0.2
matplotlib==3.8.3
multidict==6.1.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openpyxl==3.1.5
packaging==24.0
pandas==2.2.1
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.2.0
platformdirs==4.2.0
propcache==0.2.0
prov==2.0.1
psycopg2-binary==2.9.9
pycparser==2.22
pydot==3.0.2
PyJWT==2.8.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
reportlab==4.2.4
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
spyne==2.14.0
sqlparse==0.5.1
starlette==0.39.2
traits==6.3.2
truststore==0.9.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uv==0.2.25
uvicorn==0.31.0
Werkzeug==3.1.3
yarl==1.14.0
zeep==4.2.1
